third_party/blink/web_tests/fast/forms/week/week-stepup-stepdown-expected.txt

Check stepUp() and stepDown() behavior for type=week.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Invalid value
PASS stepUp("", null, null) is "1970-W02"
PASS stepDown("", null, null) is "1970-W01"
Non-number arguments
PASS stepUp("2010-W02", null, null, "0") is "2010-W02"
PASS stepDown("2010-W02", null, null, "0") is "2010-W02"
PASS stepUp("2010-W02", null, null, "foo") is "2010-W02"
PASS stepDown("2010-W02", null, null, "foo") is "2010-W02"
PASS stepUp("2010-W02", null, null, null) is "2010-W02"
PASS stepDown("2010-W02", null, null, null) is "2010-W02"
Normal cases
PASS stepUp("2010-W02", null, null) is "2010-W03"
PASS stepDown("2010-W02", null, null) is "2010-W01"
PASS stepUp("2010-W02", null, null, 10) is "2010-W12"
PASS stepDown("2010-W02", null, null, 11) is "2009-W44"
PASS stepUp("1970-W01", "4", null, 2) is "1970-W09"
PASS stepDown("1970-W01", "4", null, 3) is "1969-W41"
Step=any
PASS stepUp("2010-W02", "any", null) threw exception InvalidStateError: Failed to execute 'stepUp' on 'HTMLInputElement': This form element does not have an allowed value step..
PASS stepDown("2010-W02", "any", null) threw exception InvalidStateError: Failed to execute 'stepDown' on 'HTMLInputElement': This form element does not have an allowed value step..
Overflow/underflow
PASS stepUp("2010-W02", "3.40282346e+38", null) is "1970-W01"
PASS stepDown("2010-W02", "3.40282346e+38", null) is "1970-W01"
PASS stepUp("2010-W02", "1", "2010-W02") is "2010-W02"
PASS stepDown("2010-W02", "1", "2010-W02") is "2010-W02"


PASS successfullyParsed is true

TEST COMPLETE
